ml_model/requirements.txt
# Machine Learning Dependencies for Satellite Change Detection
# Compatible with Python 3.13
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
xgboost>=2.0.3
tensorflow>=2.15.0
matplotlib>=3.8.0
seaborn>=0.13.0
jupyter>=1.0.0
joblib>=1.4.0
requests>=2.31.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0